Issue #1796. Rename find-first-upload-key and match-upload page utils
